% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_week.R
\name{fetch_week}
\alias{fetch_week}
\title{Fetch weekly menus from the ETH Zurich Mensa Polyterrasse}
\usage{
fetch_week(headless = FALSE, timeout = 10)
}
\arguments{
\item{headless}{Logical. If TRUE, runs the browser in headless mode.}

\item{timeout}{Numeric. Timeout (in seconds) for browser interactions.}
}
\value{
A data.frame with one row per menu item and the following columns:
\describe{
\item{day}{Day of the week (e.g. "Mo", "Di").}
\item{type}{Menu line or category (e.g. "VEGAN", "HOME").}
\item{meal}{Whether it is the lunch or dinner menu for the day}
\item{title}{Title of the menu item.}
\item{description}{Menu description.}
\item{price}{Price as shown on the website.}
}
}
\description{
This function uses the \pkg{selenider} package
to open the Mensa Polyterrasse website, navigate to the weekly menu view,
and scrape all menus from Monday to Friday.
}
\details{
The returned data is unprocessed and intended to be cleaned using
\code{clean_week()}.
}
\examples{
fetch_week(headless = TRUE)

fetch_week(headless = FALSE, timeout =5)
}
\seealso{
\code{\link{clean_week}}, \code{\link{han_hunger}}
}
