% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/han_hunger.R
\name{han_hunger}
\alias{han_hunger}
\title{Select a menu from the Mensa Polyterrasse}
\usage{
han_hunger(
  df,
  day,
  time = "lunch",
  style = NULL,
  avoid = NULL,
  max_price = NULL,
  fallback = "cheapest"
)
}
\arguments{
\item{df}{A cleaned data.frame as returned by
\code{clean_week()}.}

\item{day}{Character string specifying the day of the week
(e.g. "Mo", "Di", "Mi").}

\item{time}{Character string specifying the meal time.
Either \code{"lunch"} or \code{"dinner"}. Default is lunch}

\item{style}{Optional character vector specifying allowed menu styles
(\code{c("VEGAN", "HOME", "GARDEN", "STREET")}).}

\item{avoid}{Optional character vector of ingredients or keywords to
exclude based on the menu description.}

\item{max_price}{Optional numeric value specifying the maximum allowed
student price (CHF). It has to be at least 7 when using it.}

\item{fallback}{Character string specifying the fallback strategy if no
menu matches the filters. Either \code{"cheapest"} or \code{"any"}. Default
is cheapest}
}
\value{
A one-row data.frame containing the selected menu.
If no menu can be selected, an empty data.frame is returned.
}
\description{
This function selects a single menu item from the weekly menu data
of the ETH Zurich Mensa Polyterrasse based on user-defined preferences
such as day, meal time, dietary style, excluded ingredients, and price.
}
\details{
If multiple menus match the criteria, one is selected at random.
If no menu matches, a fallback strategy is applied.

The fallback strategy is applied only if no menu matches all filters.
In this case, filtering is reduced to the \code{avoid} criterion only.
}
\examples{
\dontrun{
df_clean <- clean_week(df_raw)

han_hunger(
  df_clean,
  day = "Mi",
  time = "lunch",
  style = "HOME",
  avoid = c("Champignons"),
  max_price = 10
)
}

}
\seealso{
\code{\link{clean_week}}
}
