% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_week.R
\name{clean_week}
\alias{clean_week}
\title{Clean weekly menu data from Mensa Polyterrasse}
\usage{
clean_week(df)
}
\arguments{
\item{df}{A data.frame with the weekly ETH Mensa Menu with the Columns: day, style, title, description and price.
An example Code on how to scrape the ETH Polyterrasse website for the weekly menu is found in the README.Rmd file}
}
\value{
A cleaned data.frame which adds the column meal to indicate whether it is from the lunch or dinner menu.
}
\description{
This function cleans the raw menu data.
It removes irrelevant menu types, extracts the student price, determines
whether a menu is served for lunch or dinner, and prepares the data for
further filtering and selection.
}
\details{
Only the student price is retained. Menus labeled as soups or buffet items
are removed.
}
\examples{
\dontrun{
df_clean <- clean_week(df_raw)
}
}
\seealso{
\code{\link{han_hunger}}
}
